Rename project to `ralphs-little-helpers`
`ember-tb-test-helpers` is an awful name.

Some conversations held internal to thoughtbot yielded
`ralphs-little-helpers`, which was well received.

Our only concern is that this `ember-addon` isn't prefixed with
`ember-`.

We're not very concerned with that fact, but it's on our radar. If we
run into issues in the future, we'll pick a better name.

Amazingly, `ember-tb-test-helpers` and `ralphs-little-helpers` contain
the same number of characters :tada:.